+++ libs/qt/source/ftrack_qt/widgets/asset/asset_version_creation_widget.py
@@ -4,7 +4,7 @@
 try:
     from PySide6 import QtWidgets, QtCore
 except ImportError:
-    from PySide2 import QtWidgets, QtCore, QtGui
+    from PySide2 import QtWidgets, QtCore
 
 from ftrack_qt.widgets.thumbnails import AssetVersionThumbnail