+++ libs/qt/source/ftrack_qt/widgets/models/table_model.py
@@ -4,7 +4,7 @@
 try:
     from PySide6 import QtWidgets, QtCore
 except ImportError:
-    from PySide2 import QtWidgets, QtCore, QtGui
+    from PySide2 import QtWidgets, QtCore
 
 
 class TableModel(QtCore.QAbstractTableModel):