+++ libs/qt/source/ftrack_qt/widgets/user/ftrack_user.py
@@ -2,9 +2,9 @@
 # :copyright: Copyright (c) 2024 ftrack
 # TODO: Clean this code
 try:
-    from PySide6 import QtWidgets, QtCore, QtGui
+    from PySide6 import QtWidgets, QtCore
 except ImportError:
-    from PySide2 import QtWidgets, QtCore, QtGui
+    from PySide2 import QtWidgets, QtCore
 
 from ftrack_qt.widgets.thumbnails import UserThumbnail as UserThumbnail