+++ projects/connect-timetracker-widget/connect-plugin/hook/time_tracker_widget.py
@@ -6,9 +6,9 @@ import sys
 import operator
 
 try:
-    from PySide6 import QtWidgets, QtCore, QtGui
+    from PySide6 import QtWidgets, QtCore
 except ImportError:
-    from PySide2 import QtWidgets, QtCore, QtGui
+    from PySide2 import QtWidgets, QtCore
 
 import ftrack_api